Move process function to lib.rs
